requirements.txt
asgiref==3.8.1
contourpy==1.3.2
cycler==0.12.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
fonttools==4.58.0
joblib==1.4.2
kiwisolver==1.4.8
kmodes==0.12.2
matplotlib==3.10.3
networkx==3.4.2
numpy==2.2.4
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psycopg2==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
